+++ libs/hsm/tests/defer.c
@@ -32,7 +32,6 @@
 #include "common/compiler.h"
 #include "common/macros.h"
 #include "common/types.h"
-#include "event/event_async.h"
 #include "event/event_common.h"
 #include "event/event_pool.h"
 #include "event/event_queue.h"
@@ -189,8 +188,6 @@ static void test_defer(void) {
         AM_ASSERT(2 == am_event_alloc_get_nblocks(&alloc, /*index=*/0));
         AM_ASSERT(2 == am_event_alloc_get_nfree(&alloc, /*index=*/0));
     }
-
-    am_event_async_global_init(/*sub=*/NULL, /*nsub=*/0, &alloc);
 
     defer_init(defer_hsm_log, &alloc);